+++ CHANGELOG.md
@@ -2,8 +2,17 @@
 All notable changes will be tracked in this file
 
 ##[Unreleased]
+###Added
+- Function newBrick to BrickGeneratorOperator 
+- WallGeneratorOperator
+- Add Brick button to Add menu
+
 ###Changed
 - randVertsBMesh made a function of BrickGeneratorOperator
+- moved BrickGeneratorOperator execute code to newBrick function
+
+###Removed
+- References to original Brick class
 
 ##[0.2] - 2017-01-24
 ###Added